# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Franchise Development OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      operationId: listTasks
      summary: List tasks (optionally filtered by assignee or project)
      parameters:
        - name: assigneeId
          in: query
          required: false
          schema:
            type: string
        - name: projectId
          in: query
          required: false
          schema:
            type: string
        - name: mine
          in: query
          required: false
          schema:
            type: boolean
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskStatus"
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      tags: [Tasks]
      operationId: createTask
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      tags: [Tasks]
      operationId: getTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    patch:
      tags: [Tasks]
      operationId: updateTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      tags: [Tasks]
      operationId: deleteTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tasks/{id}/escalate:
    post:
      tags: [Tasks]
      operationId: escalateRoadblock
      summary: Escalate a roadblock on a task upward to the responsible higher-up
      description: >-
        Flags the task as Blocked and opens an escalation routed automatically to
        the executive owner of the task's initiative (with an admin/vp fallback).
        Members may only escalate their own tasks and can never choose the target.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscalateInput"
      responses:
        "200":
          description: Task with the escalation opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid input or no higher-up available to escalate to
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden — you can only escalate your own task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /blocked-tasks:
    get:
      tags: [Tasks]
      operationId: listBlockedTasks
      summary: List all open Blocked tasks across the team (excluding active escalations)
      description: >-
        Returns every open task currently in Blocked status across the team,
        excluding tasks that have an active, unresolved escalation (those appear
        in the escalation queue instead, so the two lists never duplicate a card).
        Read-only and restricted to users holding the approve-completion
        permission, mirroring the Roadblocks page gate.
      responses:
        "200":
          description: Open blocked tasks across the team
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "403":
          description: Forbidden — requires the approve-completion permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /escalations/{id}/resolve:
    post:
      tags: [Tasks]
      operationId: resolveEscalation
      summary: Resolve a roadblock escalation (returns to original developer by default)
      description: >-
        Clears the Blocked status and records who resolved it and when. By default
        the task returns to the original developer; the higher-up may instead
        reassign it to someone else. Only the higher-up the escalation was routed
        to (or an admin/vp) may resolve.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveEscalationInput"
      responses:
        "200":
          description: Task with the escalation resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid input or no active escalation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden — only the higher-up may resolve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /tasks/bulk-assign:
    post:
      tags: [Tasks]
      operationId: bulkAssignTasks
      summary: Assign a task to a group — creates a parent task plus one child per group member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAssignInput"
      responses:
        "201":
          description: Created tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkAssignResult"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    TaskStatus:
      type: string
      enum: [Not Started, In Progress, Waiting on Someone, Blocked, Pending Review, Complete]
    TaskRole:
      type: string
      enum: [Owner, Contributor, Support]
...
    Task:
      type: object
      required:
        - id
        - projectId
        - assigneeId
        - role
        - description
        - priority
        - status
        - percentComplete
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        projectId:
          type: string
        projectName:
          type: string
        projectCode:
          type: string
        focusAreaId:
          type: integer
        focusAreaName:
          type: string
        assigneeId:
          type: string
        assigneeName:
          type: string
        assigneeAvatarColor:
          type: string
        role:
          $ref: "#/components/schemas/TaskRole"
        description:
          type: string
        priority:
          $ref: "#/components/schemas/TaskPriority"
        status:
          $ref: "#/components/schemas/TaskStatus"
        percentComplete:
          type: integer
          minimum: 0
          maximum: 100
        startDate:
          type: [string, 'null']
          format: date
        dueDate:
          type: [string, 'null']
          format: date
        dependency:
          type: [string, 'null']
        roadblock:
          type: [string, 'null']
        lastUpdate:
          type: [string, 'null']
        nextStep:
          type: [string, 'null']
        blockedById:
          type: [string, 'null']
        blockedByName:
          type: [string, 'null']
        blockNotes:
          type: [string, 'null']
        escalatedToId:
          type: [string, 'null']
        escalatedToName:
          type: [string, 'null']
        escalatedById:
          type: [string, 'null']
        escalatedByName:
          type: [string, 'null']
        escalationOriginalAssigneeId:
          type: [string, 'null']
        escalationOriginalAssigneeName:
          type: [string, 'null']
        escalatedAt:
          type: [string, 'null']
          format: date-time
        escalationResolvedById:
          type: [string, 'null']
        escalationResolvedByName:
          type: [string, 'null']
        escalationResolvedAt:
          type: [string, 'null']
          format: date-time
        escalationResolutionNote:
          type: [string, 'null']
        isEscalated:
          type: boolean
        groupId:
          type: [integer, 'null']
        parentTaskId:
          type: [integer, 'null']
        isGroupParent:
          type: boolean
        isOverdue:
          type: boolean
        isDueSoon:
          type: boolean
        isBlockingViewer:
          type: boolean
          description: >-
            True when the requesting user is the recorded blocker of this
            currently-blocked task. Lets the personal task list flag items that
            appear because someone is waiting on the viewer to unblock them.
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/TaskAttachment"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    TaskInput:
      type: object
      required:
        - projectId
        - assigneeId
        - role
        - description
        - priority
        - status
        - percentComplete
      properties:
        projectId:
          type: string
          minLength: 1
        assigneeId:
          type: string
          minLength: 1
        groupId:
          type: [integer, 'null']
        parentTaskId:
          type: [integer, 'null']
        role:
          $ref: "#/components/schemas/TaskRole"
        description:
          type: string
          minLength: 1
        priority:
          $ref: "#/components/schemas/TaskPriority"
        status:
          $ref: "#/components/schemas/TaskStatus"
        percentComplete:
          type: integer
          minimum: 0
          maximum: 100
        startDate:
          type: [string, 'null']
          format: date
        dueDate:
          type: [string, 'null']
          format: date
        dependency:
          type: [string, 'null']
        roadblock:
          type: [string, 'null']
        lastUpdate:
          type: [string, 'null']
        nextStep:
          type: [string, 'null']
        blockedById:
          type: [string, 'null']
        blockNotes:
          type: [string, 'null']
    TaskUpdate:
      type: object
      properties:
        projectId:
          type: string
        assigneeId:
          type: string
        role:
          $ref: "#/components/schemas/TaskRole"
        description:
          type: string
        priority:
          $ref: "#/components/schemas/TaskPriority"
        status:
          $ref: "#/components/schemas/TaskStatus"
        percentComplete:
          type: integer
          minimum: 0
          maximum: 100
        startDate:
          type: [string, 'null']
          format: date
        dueDate:
          type: [string, 'null']
          format: date
        dependency:
          type: [string, 'null']
        roadblock:
          type: [string, 'null']
        lastUpdate:
          type: [string, 'null']
        nextStep:
          type: [string, 'null']
        blockedById:
          type: [string, 'null']
        blockNotes:
          type: [string, 'null']
    EscalateInput:
      type: object
      required: [note]
      properties:
        note:
          type: string
          minLength: 1
          description: What is blocking the developer, sent to the higher-up.
...